Keep GoPeak's packaging and docs aligned with the current MCP surface
This folds the three highest-priority cleanup items into one reviewable delivery: tool metadata and group constants move out of the 9.8k-line entrypoint, shell-hook installation becomes explicitly opt-in during postinstall, and repository-facing docs/metadata now match the verified prompt/resource/tool behavior in v2.3.4.

The goal was to reduce maintenance drag without changing the actual MCP/runtime behavior that current users rely on. The implementation stays dependency-free, keeps compact/full profile semantics intact, and adds regression coverage for package/server metadata plus the new postinstall contract.

Constraint: Preserve existing MCP behavior while reducing entrypoint sprawl and install-side effects
Constraint: No new dependencies for packaging, testing, or refactoring
Rejected: Full multi-file server handler split in one PR | too broad for a safe follow-up change
Rejected: Remove shell-hook support entirely | explicit opt-in keeps the feature without surprising install mutations
Confidence: medium
Scope-risk: moderate
Reversibility: clean
Directive: Keep README/package/server metadata and install-flow docs in sync whenever tool counts, MCP capabilities, or setup behavior change
Tested: npm run test:setup
Tested: npm run ci
Tested: npm run test:dynamic-groups
Tested: npm run test:integration
Not-tested: Manual verification of every translated README shell-hook paragraph beyond the updated count references
