+++ README-zh.md
@@ -542,7 +542,7 @@ MIT 许可证 - 详情请参阅 [LICENSE](LICENSE)。
 
 ## 统计数据
 
-- **95+ 工具** — 涵盖场景管理、脚本、资源、动画、配置、调试、截图、输入注入、LSP、DAP 和资源管理的综合工具
+- **110+ 工具** — 涵盖场景管理、脚本、资源、动画、配置、调试、截图、输入注入、LSP、DAP 和资源管理的综合工具
 - **MCP 资源** — 用于直接项目文件访问的 `godot://` URI 协议
 - **GDScript LSP** — 通过 Godot 语言服务器提供实时诊断、补全、悬停和符号
 - **调试适配器 (DAP)** — 断点、步进、堆栈跟踪和控制台输出捕获
