+++ server.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://static.modelcontextprotocol.io/schemas/2025-12-11/server.schema.json",
   "name": "io.github.HaD0Yun/gopeak",
-  "description": "GoPeak — The most comprehensive MCP server for Godot Engine. 95+ tools, LSP, DAP, screenshots.",
+  "description": "GoPeak \u2014 MCP server for Godot Engine with 110+ tools, compact/dynamic profiles, LSP, DAP, screenshots, and runtime automation.",
   "repository": {
     "url": "https://github.com/HaD0Yun/Gopeak-godot-mcp",
     "source": "github"
